chore(graphql): modularize resolvers and typedefs
- modularized resolvers and typedefs into separate files
  - without doing this the typedefs will grow to an unmanagable size with oparl
