+++ src/index.ts
@@ -1,5 +1,5 @@
 import express from 'express';
-import { ApolloServer } from 'apollo-server-express';
+import { ApolloServer, makeExecutableSchema } from 'apollo-server-express';
 import mongoose from 'mongoose';
 
 import { resolvers } from './resolvers';
@@ -7,10 +7,11 @@ import { typeDefs } from './typeDefs';
 
 const startServer = async () => {
   const app = express();
-
   const server = new ApolloServer({
-    typeDefs,
-    resolvers,
+    schema: makeExecutableSchema({
+      typeDefs,
+      resolvers,
+    }),
   });
 
   server.applyMiddleware({ app });
