+++ src/typeDefs/date.ts
@@ -1,0 +1,5 @@
+import { gql } from 'apollo-server-core';
+
+export const dateGQL = gql`
+  scalar Date
+`;
