+++ src/typeDefs/location.ts
@@ -1,0 +1,8 @@
+import { gql } from 'apollo-server-core';
+
+export const locationGQL = gql`
+  type Location {
+    lat: String
+    lon: String
+  }
+`;
