+++ src/typeDefs/roadwork.ts
@@ -1,0 +1,9 @@
+import { gql } from 'apollo-server-core';
+
+export const roadworkGQL = gql`
+  type Roadwork {
+    id: ID!
+    name: String!
+    description: String
+  }
+`;
